Refactor Makefile and improve error checking
Prepare the Makefile for multiple folders.

Also improve skip_ulp_redirect_insns to check if the function to be
redirected is actually a livepatched function. Glibc has too many
functions written in plain asm (for instance, getpid which moves
a value to rax and then execute syscall), and that would catch such
cases when developing the livepatches.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>