Use malloc() as new livepatch info instead of strdup(NULL)
Instead of using strdup(NULL) as a way to test if livepatch was applied,
we now return this information in malloc. This way if the livepatch
isn't applied the test does not crash.

Also be careful about the order in which the livepatches are applied.
We must now apply the libc_livepatch last because it is used to
trigger the test continue.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>