Add necessary barriers after code generation for ppc64le
powerpc64le is a weakly-ordered architecture, which means barriers are
necessary to synchronize some things in the CPU.  In this case since
we are writing code into the memory, we need make sure the i-cache
is synchronized with the d-cache.

Hence the asm instructions inserted.

Source: kernel eBPF:
arch/powerpc/lib/code-patching.c
  __patch_mem

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>